Fix type of string lengths.
Those values are used in various call (including zpp) where
a size_t is expected.

This fix 82 failed test on ppc64
(bigendian is perfect to detect stack issue)

+++ ext/spl/spl_directory.h
@@ -63,10 +63,10 @@ struct _spl_filesystem_object {
 	void               *oth;
 	spl_other_handler  *oth_handler;
 	char               *_path;
-	int                _path_len;
+	size_t             _path_len;
 	char               *orig_path;
 	char               *file_name;
-	int                file_name_len;
+	size_t             file_name_len;
 	SPL_FS_OBJ_TYPE    type;
 	zend_long               flags;
 	zend_class_entry   *file_class;
@@ -76,7 +76,7 @@ struct _spl_filesystem_object {
 			php_stream         *dirp;
 			php_stream_dirent  entry;
 			char               *sub_path;
-			int                sub_path_len;
+			size_t             sub_path_len;
 			int                index;
 			int                is_recursive;
 			zend_function      *func_rewind;
@@ -88,7 +88,7 @@ struct _spl_filesystem_object {
 			php_stream_context *context;
 			zval               *zcontext;
 			char               *open_mode;
-			int                open_mode_len;
+			size_t             open_mode_len;
 			zval               current_zval;
 			char               *current_line;
 			size_t             current_line_len;
